Corrections mineures dans la synthèse des images.

+++ audela/src/audela/libtt/src/tt_user5.c
@@ -222,8 +222,8 @@ int tt_ima_rot(TT_IMA_SERIES *pseries)
 #define max(a,b) (a>b?a:b)
 #endif
 
-   w = max( fabs(p_in->naxis1*cos_theta - p_in->naxis2*sin_theta) , fabs(p_in->naxis1*cos_theta + p_in->naxis2*sin_theta) );
-   h = max( fabs(p_in->naxis1*sin_theta + p_in->naxis2*cos_theta) , fabs(p_in->naxis1*sin_theta - p_in->naxis2*cos_theta) );
+   w = (int)max( fabs(p_in->naxis1*cos_theta - p_in->naxis2*sin_theta) , fabs(p_in->naxis1*cos_theta + p_in->naxis2*sin_theta) );
+   h = (int)max( fabs(p_in->naxis1*sin_theta + p_in->naxis2*cos_theta) , fabs(p_in->naxis1*sin_theta - p_in->naxis2*cos_theta) );
    
    /* --- coordonnéees du centre de rotation dans l'image initiale --- */
    x0=pseries->user5.x0;
